[MODCONSKC-62] Split sync and create primary affiliation services
