+++ src/main/java/org/folio/consortia/service/SyncPrimaryAffiliationService.java
@@ -16,11 +16,11 @@ public interface SyncPrimaryAffiliationService {
   void syncPrimaryAffiliations(UUID consortiumId, String tenantId, String centralTenantId);
 
   /**
-   * Create affiliations between central tenant and primary users in the tenant specified in the body
+   * Sync primary affiliation for specified users in the specified tenant
    *
    * @param consortiumId               - consortium unique identifier
    * @param centralTenantId            - central tenant unique identifier
    * @param syncPrimaryAffiliationBody - object with tenant id and user list
    */
-  void createPrimaryUserAffiliations(UUID consortiumId, String centralTenantId, SyncPrimaryAffiliationBody syncPrimaryAffiliationBody);
+  void syncPrimaryUserAffiliations(UUID consortiumId, String centralTenantId, SyncPrimaryAffiliationBody syncPrimaryAffiliationBody);
 }
